fix: upgrade the behaviors of testng
