+++ fhir-audit/src/test/java/testng.xml
@@ -1,4 +1,4 @@
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
 <suite name="FHIRAuditCadfTestSuite">
     <test name="Tests">
         <classes>
